+++ gsa/src/web/pages/schedules/dialog.js
@@ -328,7 +328,10 @@ class ScheduleDialog extends React.Component {
         interval: isPreDefined ? 1 : interval,
         monthdays: setMonthydays ? monthdays : undefined,
         weekdays: setWeekdays ? weekdays : undefined,
-        summary: name,
+        // convert name to string explicitly to not run into:
+        // `TypeError: e.replace is not a function`
+        // when name is just numbers.
+        summary: name.toString(),
         startDate,
       },
       timezone,